Update .env example to add grace period for JWT
I spent 6 hours trying to get this working on an apache install. When trying to log in it returned an error saying the token was on the blacklist. If the old token and new token were created in the same second, they were both being added to the blacklist due to this bug: https://github.com/tymondesigns/jwt-auth/issues/83

Adding the grace period in the .env will hopefully stop someone else struggling with this issue.